Generate data (random variable) from a given distribution

The parameters of the expected random variable are used to draw a random sample from the given probability distribution.

Usage

generate_data(
  distribution,
  n,
  mean = NULL,
  sd = NULL,
  size = NULL,
  prob = NULL
)

Arguments

distribution

The probability distribution.

n

The number of values to draw from the probability distribution.

mean

The mean (central tendency).

sd

The standard deviation (variability).

size

The target number of successful trials.

prob

The probability of success in each independent and identical trial.

Value

A vector of random deviates whose length is n

Examples

generate_data("normal", 6, 2.5, 1.3)
#> [1]  0.6799434  2.8319122 -0.6684427  2.4927573  3.3080185  3.9929351
generate_data("poisson", 6, 2.5)
#> [1] 0 2 2 1 2 0
generate_data("logistic", 6, prob = 0.5, size = 1)
#> [1] 0 1 0 1 1 0
generate_data("negbin", 6, 2.5, size = 1)
#> [1] 3 2 3 1 0 3